move Registry tests, plots_test.js -> register_test.js
