+++ test/jasmine/tests/plots_test.js
@@ -245,140 +245,6 @@ describe('Test Plots', function() {
         });
     });
 
-    describe('Plots.register, getModule, and traceIs', function() {
-        beforeEach(function() {
-            this.modulesKeys = Object.keys(Plots.modules);
-            this.allTypesKeys = Object.keys(Plots.allTypes);
-            this.allCategoriesKeys = Object.keys(Plots.allCategories);
-
-            this.fakeModule = {
-                calc: function() { return 42; },
-                plot: function() { return 1000000; }
-            };
-            this.fakeModule2 = {
-                plot: function() { throw new Error('nope!'); }
-            };
-
-            Plots.register(this.fakeModule, 'newtype', ['red', 'green']);
-
-            spyOn(console, 'warn');
-        });
-
-        afterEach(function() {
-            function revertObj(obj, initialKeys) {
-                Object.keys(obj).forEach(function(k) {
-                    if(initialKeys.indexOf(k) === -1) delete obj[k];
-                });
-            }
-
-            revertObj(Plots.modules, this.modulesKeys);
-            revertObj(Plots.allTypes, this.allTypesKeys);
-            revertObj(Plots.allCategories, this.allCategoriesKeys);
-        });
-
-        it('should not reregister a type', function() {
-            Plots.register(this.fakeModule2, 'newtype', ['yellow', 'blue']);
-            expect(Plots.allCategories.yellow).toBeUndefined();
-        });
-
-        it('should find the module for a type', function() {
-            expect(Plots.getModule('newtype')).toBe(this.fakeModule);
-            expect(Plots.getModule({type: 'newtype'})).toBe(this.fakeModule);
-        });
-
-        it('should return false for types it doesn\'t know', function() {
-            expect(Plots.getModule('notatype')).toBe(false);
-            expect(Plots.getModule({type: 'notatype'})).toBe(false);
-            expect(Plots.getModule({type: 'newtype', r: 'this is polar'})).toBe(false);
-        });
-
-        it('should find the categories for this type', function() {
-            expect(Plots.traceIs('newtype', 'red')).toBe(true);
-            expect(Plots.traceIs({type: 'newtype'}, 'red')).toBe(true);
-        });
-
-        it('should not find other real categories', function() {
-            expect(Plots.traceIs('newtype', 'cartesian')).toBe(false);
-            expect(Plots.traceIs({type: 'newtype'}, 'cartesian')).toBe(false);
-            expect(console.warn).not.toHaveBeenCalled();
-        });
-    });
-
-    describe('Plots.registerSubplot', function() {
-        var fake = {
-            name: 'fake',
-            attr: 'abc',
-            idRoot: 'cba',
-            attrRegex: /^abc([2-9]|[1-9][0-9]+)?$/,
-            idRegex: /^cba([2-9]|[1-9][0-9]+)?$/,
-            attributes: { stuff: { 'more stuff': 102102 } }
-        };
-
-        Plots.registerSubplot(fake);
-
-        var subplotsRegistry = Plots.subplotsRegistry;
-
-        it('should register attr, idRoot and attributes', function() {
-            expect(subplotsRegistry.fake.attr).toEqual('abc');
-            expect(subplotsRegistry.fake.idRoot).toEqual('cba');
-            expect(subplotsRegistry.fake.attributes)
-                .toEqual({stuff: { 'more stuff': 102102 }});
-        });
-
-        describe('registered subplot type attribute regex', function() {
-            it('should compile to correct attribute regex string', function() {
-                expect(subplotsRegistry.fake.attrRegex.toString())
-                    .toEqual('/^abc([2-9]|[1-9][0-9]+)?$/');
-            });
-
-            var shouldPass = [
-                'abc', 'abc2', 'abc3', 'abc10', 'abc9', 'abc100', 'abc2002'
-            ];
-            var shouldFail = [
-                '0abc', 'abc0', 'abc1', 'abc021321', 'abc00021321'
-            ];
-
-            shouldPass.forEach(function(s) {
-                it('considers ' + JSON.stringify(s) + 'as a correct attribute name', function() {
-                    expect(subplotsRegistry.fake.attrRegex.test(s)).toBe(true);
-                });
-            });
-
-            shouldFail.forEach(function(s) {
-                it('considers ' + JSON.stringify(s) + 'as an incorrect attribute name', function() {
-                    expect(subplotsRegistry.fake.attrRegex.test(s)).toBe(false);
-                });
-            });
-        });
-
-        describe('registered subplot type id regex', function() {
-            it('should compile to correct id regular expression', function() {
-                expect(subplotsRegistry.fake.idRegex.toString())
-                    .toEqual('/^cba([2-9]|[1-9][0-9]+)?$/');
-            });
-
-            var shouldPass = [
-                'cba', 'cba2', 'cba3', 'cba10', 'cba9', 'cba100', 'cba2002'
-            ];
-            var shouldFail = [
-                '0cba', 'cba0', 'cba1', 'cba021321', 'cba00021321'
-            ];
-
-            shouldPass.forEach(function(s) {
-                it('considers ' + JSON.stringify(s) + 'as a correct attribute name', function() {
-                    expect(subplotsRegistry.fake.idRegex.test(s)).toBe(true);
-                });
-            });
-
-            shouldFail.forEach(function(s) {
-                it('considers ' + JSON.stringify(s) + 'as an incorrect attribute name', function() {
-                    expect(subplotsRegistry.fake.idRegex.test(s)).toBe(false);
-                });
-            });
-        });
-
-    });
-
     describe('Plots.resize', function() {
         var gd;
