+++ test/jasmine/tests/register_test.js
@@ -1,15 +1,152 @@
 var Plotly = require('@lib/index');
+var Registry = require('@src/registry');
+
+describe('Test Registry', function() {
+    'use strict';
+
+    describe('register, getModule, and traceIs', function() {
+        beforeEach(function() {
+            this.modulesKeys = Object.keys(Registry.modules);
+            this.allTypesKeys = Object.keys(Registry.allTypes);
+            this.allCategoriesKeys = Object.keys(Registry.allCategories);
+
+            this.fakeModule = {
+                calc: function() { return 42; },
+                plot: function() { return 1000000; }
+            };
+            this.fakeModule2 = {
+                plot: function() { throw new Error('nope!'); }
+            };
+
+            Registry.register(this.fakeModule, 'newtype', ['red', 'green']);
+
+            spyOn(console, 'warn');
+        });
+
+        afterEach(function() {
+            function revertObj(obj, initialKeys) {
+                Object.keys(obj).forEach(function(k) {
+                    if(initialKeys.indexOf(k) === -1) delete obj[k];
+                });
+            }
+
+            revertObj(Registry.modules, this.modulesKeys);
+            revertObj(Registry.allTypes, this.allTypesKeys);
+            revertObj(Registry.allCategories, this.allCategoriesKeys);
+        });
+
+        it('should not reregister a type', function() {
+            Registry.register(this.fakeModule2, 'newtype', ['yellow', 'blue']);
+            expect(Registry.allCategories.yellow).toBeUndefined();
+        });
+
+        it('should find the module for a type', function() {
+            expect(Registry.getModule('newtype')).toBe(this.fakeModule);
+            expect(Registry.getModule({type: 'newtype'})).toBe(this.fakeModule);
+        });
+
+        it('should return false for types it doesn\'t know', function() {
+            expect(Registry.getModule('notatype')).toBe(false);
+            expect(Registry.getModule({type: 'notatype'})).toBe(false);
+            expect(Registry.getModule({type: 'newtype', r: 'this is polar'})).toBe(false);
+        });
+
+        it('should find the categories for this type', function() {
+            expect(Registry.traceIs('newtype', 'red')).toBe(true);
+            expect(Registry.traceIs({type: 'newtype'}, 'red')).toBe(true);
+        });
+
+        it('should not find other real categories', function() {
+            expect(Registry.traceIs('newtype', 'cartesian')).toBe(false);
+            expect(Registry.traceIs({type: 'newtype'}, 'cartesian')).toBe(false);
+            expect(console.warn).not.toHaveBeenCalled();
+        });
+    });
+
+    describe('Registry.registerSubplot', function() {
+        var fake = {
+            name: 'fake',
+            attr: 'abc',
+            idRoot: 'cba',
+            attrRegex: /^abc([2-9]|[1-9][0-9]+)?$/,
+            idRegex: /^cba([2-9]|[1-9][0-9]+)?$/,
+            attributes: { stuff: { 'more stuff': 102102 } }
+        };
+
+        Registry.registerSubplot(fake);
+
+        var subplotsRegistry = Registry.subplotsRegistry;
+
+        it('should register attr, idRoot and attributes', function() {
+            expect(subplotsRegistry.fake.attr).toEqual('abc');
+            expect(subplotsRegistry.fake.idRoot).toEqual('cba');
+            expect(subplotsRegistry.fake.attributes)
+                .toEqual({stuff: { 'more stuff': 102102 }});
+        });
+
+        describe('registered subplot type attribute regex', function() {
+            it('should compile to correct attribute regex string', function() {
+                expect(subplotsRegistry.fake.attrRegex.toString())
+                    .toEqual('/^abc([2-9]|[1-9][0-9]+)?$/');
+            });
+
+            var shouldPass = [
+                'abc', 'abc2', 'abc3', 'abc10', 'abc9', 'abc100', 'abc2002'
+            ];
+            var shouldFail = [
+                '0abc', 'abc0', 'abc1', 'abc021321', 'abc00021321'
+            ];
+
+            shouldPass.forEach(function(s) {
+                it('considers ' + JSON.stringify(s) + 'as a correct attribute name', function() {
+                    expect(subplotsRegistry.fake.attrRegex.test(s)).toBe(true);
+                });
+            });
+
+            shouldFail.forEach(function(s) {
+                it('considers ' + JSON.stringify(s) + 'as an incorrect attribute name', function() {
+                    expect(subplotsRegistry.fake.attrRegex.test(s)).toBe(false);
+                });
+            });
+        });
+
+        describe('registered subplot type id regex', function() {
+            it('should compile to correct id regular expression', function() {
+                expect(subplotsRegistry.fake.idRegex.toString())
+                    .toEqual('/^cba([2-9]|[1-9][0-9]+)?$/');
+            });
+
+            var shouldPass = [
+                'cba', 'cba2', 'cba3', 'cba10', 'cba9', 'cba100', 'cba2002'
+            ];
+            var shouldFail = [
+                '0cba', 'cba0', 'cba1', 'cba021321', 'cba00021321'
+            ];
+
+            shouldPass.forEach(function(s) {
+                it('considers ' + JSON.stringify(s) + 'as a correct attribute name', function() {
+                    expect(subplotsRegistry.fake.idRegex.test(s)).toBe(true);
+                });
+            });
+
+            shouldFail.forEach(function(s) {
+                it('considers ' + JSON.stringify(s) + 'as an incorrect attribute name', function() {
+                    expect(subplotsRegistry.fake.idRegex.test(s)).toBe(false);
+                });
+            });
+        });
+
+    });
+});
 
 describe('the register function', function() {
     'use strict';
 
-    var Plots = Plotly.Plots;
-
     beforeEach(function() {
-        this.modulesKeys = Object.keys(Plots.modules);
-        this.allTypesKeys = Object.keys(Plots.allTypes);
-        this.allCategoriesKeys = Object.keys(Plots.allCategories);
-        this.allTransformsKeys = Object.keys(Plots.transformsRegistry);
+        this.modulesKeys = Object.keys(Registry.modules);
+        this.allTypesKeys = Object.keys(Registry.allTypes);
+        this.allCategoriesKeys = Object.keys(Registry.allCategories);
+        this.allTransformsKeys = Object.keys(Registry.transformsRegistry);
     });
 
     afterEach(function() {
@@ -19,10 +156,10 @@ describe('the register function', function() {
             });
         }
 
-        revertObj(Plots.modules, this.modulesKeys);
-        revertObj(Plots.allTypes, this.allTypesKeys);
-        revertObj(Plots.allCategories, this.allCategoriesKeys);
-        revertObj(Plots.transformsRegistry, this.allTransformsKeys);
+        revertObj(Registry.modules, this.modulesKeys);
+        revertObj(Registry.allTypes, this.allTypesKeys);
+        revertObj(Registry.allCategories, this.allCategoriesKeys);
+        revertObj(Registry.transformsRegistry, this.allTransformsKeys);
     });
 
     it('should throw an error when no argument is given', function() {
@@ -44,7 +181,7 @@ describe('the register function', function() {
             Plotly.register(mockTrace1);
         }).not.toThrow();
 
-        expect(Plotly.Plots.getModule('mockTrace1')).toBe(mockTrace1);
+        expect(Registry.getModule('mockTrace1')).toBe(mockTrace1);
     });
 
     it('should work with an array of modules', function() {
@@ -60,7 +197,7 @@ describe('the register function', function() {
             Plotly.register([mockTrace2]);
         }).not.toThrow();
 
-        expect(Plotly.Plots.getModule('mockTrace2')).toBe(mockTrace2);
+        expect(Registry.getModule('mockTrace2')).toBe(mockTrace2);
     });
 
     it('should throw an error when an invalid module is given', function() {
@@ -99,6 +236,6 @@ describe('the register function', function() {
             Plotly.register(transformModule);
         }).not.toThrow();
 
-        expect(Plots.transformsRegistry['mah-transform']).toBeDefined();
+        expect(Registry.transformsRegistry['mah-transform']).toBeDefined();
     });
 });
